Leeang changes

+++ src/http.js
@@ -21,18 +21,21 @@ instance.interceptors.request.use(function (config) {
 });
 
 // Add a response interceptor
-instance.interceptors.response.use(function (response) {
-    // Any status code that lie within the range of 2xx cause this function to trigger
-    // Do something with response data
-    return response;
-}, function (error) {
-    // Any status codes that falls outside the range of 2xx cause this function to trigger
-    // Do something with response error
-    if (error.response.status === 401 || error.response.status === 403) {
-        localStorage.clear();
-        window.location = "/login";
+instance.interceptors.response.use(
+    function (response) {
+        // Any status code that lies within the range of 2xx causes this function to trigger
+        // Do something with response data
+        return response;
+    },
+    function (error) {
+        // Any status codes that fall outside the range of 2xx cause this function to trigger
+        // Do something with response error
+        if (error.response && (error.response.status === 401 || error.response.status === 403)) {
+            localStorage.clear();
+            window.location = "/login";
+        }
+        return Promise.reject(error);
     }
-    return Promise.reject(error);
-});
+);
 
 export default instance;
